agrega: archivo defaultTags
