+++ src/modules/taskList/Tags/model/defaultTags.ts
@@ -1,0 +1,56 @@
+import { TagGroup } from "./tags";
+
+
+// Este grupo de etiquetas tiene traduccion, tenelo en cuenta al realizar cambios
+
+export const eisenhowerTagGroup: TagGroup = {
+	id: 'Eisenhower',
+	tags: [
+		{
+			id: '1',
+			name: 'Importante',
+			variant: 'purple-subtle',
+		},
+		{
+			id: '2',
+			name: 'Necesario',
+			variant: 'green-subtle',
+		},
+		{
+			id: '3',
+			name: 'Urgente',
+			variant: 'red-subtle',
+		},
+	],
+}
+
+export const devTagGroup: TagGroup = {
+		id: 'Dev',
+		tags: [
+			{
+				id: 'Importante',
+				name: 'Importante',
+				variant: 'purple-subtle',
+			},
+			{
+				id: 'Necesario',
+				name: 'Necesario',
+				variant: 'green-subtle',
+			},
+			{
+				id: 'Urgente',
+				name: 'Urgente',
+				variant: 'red-subtle',
+			},
+			{
+				id: 'Explorar',
+				name: 'Explorar',
+				variant: 'blue-subtle',
+			},
+        	{
+				id: 'Resolver',
+				name: 'Resolver',
+				variant: 'amber-subtle',
+            },
+		],
+}
